refactor(ai): extract pure model-identity + class-name helpers into vailabel-models (ai slice 1)
First slice of the ai-module migration toward emptying src/domain. Move the
pure, I/O-free helpers out of the 2,268-line AiService into vailabel-models:

- domain/class_names.rs: COCO-80 table + class-name resolution (normalize /
  extract from array|object|numeric-map / builtin family labels).
- domain/identity.rs: model extension/runtime, Ultralytics family/variant/task
  suffix, default registry rank, display version string, plural-tolerant class
  token match.

AiService now imports the 12 it still calls; the feature-gated
prediction_unsupported_reason, the file-reading parse_config_class_names, the
std::process ONNX export, registry_id_for_model (uses value_string), and all
ONNX/HTTP infrastructure stay in the binary for later slices. crates/models
stays FULLY_PURE (serde_json + std::path only). 6 crate unit tests added.

Green: cargo test -p vailabel-models (6), arch-tests (4), default cargo check
(no warnings).
