+++ apps/studio/src-tauri/crates/models/Cargo.toml
@@ -9,3 +9,4 @@ name = "vailabel_models"
 
 [dependencies]
 serde = { workspace = true }
+serde_json = { workspace = true }
