+++ apps/studio/src-tauri/src/domain/ai/service.rs
@@ -10,7 +10,7 @@ use crate::{
     now_iso, value_string, AppError,
 };
 use reqwest::blocking::Client;
-use serde_json::{json, Map, Value};
+use serde_json::{json, Value};
 use std::collections::HashMap;
 use std::fs;
 use std::io::copy;
@@ -20,92 +20,14 @@ use std::sync::{Arc, Mutex};
 use std::time::Instant;
 use tauri::Manager;
 use uuid::Uuid;
+use vailabel_models::domain::{
+    build_model_version, builtin_class_names, class_token_matches, extract_class_names_from_value,
+    infer_default_rank, infer_model_family_and_variant, infer_model_runtime, is_pytorch_checkpoint,
+    model_extension, non_detection_reason, task_suffix, task_type_for_category,
+};
 
 const APP_NAME: &str = "Vailabel Studio";
 const DEFAULT_AI_LABEL_COLOR: &str = "#22c55e";
-
-const COCO_80_CLASS_NAMES: &[&str] = &[
-    "person",
-    "bicycle",
-    "car",
-    "motorcycle",
-    "airplane",
-    "bus",
-    "train",
-    "truck",
-    "boat",
-    "traffic light",
-    "fire hydrant",
-    "stop sign",
-    "parking meter",
-    "bench",
-    "bird",
-    "cat",
-    "dog",
-    "horse",
-    "sheep",
-    "cow",
-    "elephant",
-    "bear",
-    "zebra",
-    "giraffe",
-    "backpack",
-    "umbrella",
-    "handbag",
-    "tie",
-    "suitcase",
-    "frisbee",
-    "skis",
-    "snowboard",
-    "sports ball",
-    "kite",
-    "baseball bat",
-    "baseball glove",
-    "skateboard",
-    "surfboard",
-    "tennis racket",
-    "bottle",
-    "wine glass",
-    "cup",
-    "fork",
-    "knife",
-    "spoon",
-    "bowl",
-    "banana",
-    "apple",
-    "sandwich",
-    "orange",
-    "broccoli",
-    "carrot",
-    "hot dog",
-    "pizza",
-    "donut",
-    "cake",
-    "chair",
-    "couch",
-    "potted plant",
-    "bed",
-    "dining table",
-    "toilet",
-    "tv",
-    "laptop",
-    "mouse",
-    "remote",
-    "keyboard",
-    "cell phone",
-    "microwave",
-    "oven",
-    "toaster",
-    "sink",
-    "refrigerator",
-    "book",
-    "clock",
-    "vase",
-    "scissors",
-    "teddy bear",
-    "hair drier",
-    "toothbrush",
-];
 
 /// One loaded inference engine, keyed by model id in [`AiService::engine_cache`].
 /// Holding distinct engines side by side lets a detect→segment turn keep both the
@@ -1054,70 +976,6 @@ struct ResolvedPredictionLabel {
     label_color: Option<String>,
 }
 
-fn normalize_class_names(names: Vec<String>) -> Option<Vec<String>> {
-    let normalized = names
-        .into_iter()
-        .map(|value| value.trim().to_string())
-        .filter(|value| !value.is_empty())
-        .collect::<Vec<_>>();
-
-    if normalized.is_empty() {
-        None
-    } else {
-        Some(normalized)
-    }
-}
-
-fn extract_named_value_list(entries: &[Value]) -> Option<Vec<String>> {
-    let values = entries
-        .iter()
-        .map(|value| {
-            value.as_str().map(ToString::to_string).or_else(|| {
-                value
-                    .get("name")
-                    .and_then(Value::as_str)
-                    .map(ToString::to_string)
-            })
-        })
-        .collect::<Option<Vec<String>>>()?;
-
-    normalize_class_names(values)
-}
-
-fn extract_named_value_map(entries: &Map<String, Value>) -> Option<Vec<String>> {
-    let mut numbered_entries = entries
-        .iter()
-        .filter_map(|(key, value)| Some((key.parse::<usize>().ok()?, value.as_str()?)))
-        .collect::<Vec<_>>();
-    numbered_entries.sort_by_key(|(index, _)| *index);
-
-    if numbered_entries.is_empty() {
-        return None;
-    }
-
-    normalize_class_names(
-        numbered_entries
-            .into_iter()
-            .map(|(_, value)| value.to_string())
-            .collect(),
-    )
-}
-
-fn extract_class_names_from_value(value: &Value) -> Option<Vec<String>> {
-    match value {
-        Value::Array(entries) => extract_named_value_list(entries),
-        Value::Object(entries) => {
-            for key in ["classNames", "class_names", "names", "labels", "classes"] {
-                if let Some(candidate) = entries.get(key).and_then(extract_class_names_from_value) {
-                    return Some(candidate);
-                }
-            }
-            extract_named_value_map(entries)
-        }
-        _ => None,
-    }
-}
-
 fn parse_config_class_names(config_path: &str) -> Result<Option<Vec<String>>, AppError> {
     if config_path.trim().is_empty() {
         return Ok(None);
@@ -1144,36 +1002,6 @@ fn parse_config_class_names(config_path: &str) -> Result<Option<Vec<String>>, Ap
     };
 
     Ok(extract_class_names_from_value(&parsed))
-}
-
-fn builtin_class_names(family: &str, category: &str) -> Option<Vec<String>> {
-    if category.eq_ignore_ascii_case("detection")
-        && matches!(
-            family.trim().to_ascii_lowercase().as_str(),
-            "yolo26" | "yolo11" | "yolov8"
-        )
-    {
-        return Some(
-            COCO_80_CLASS_NAMES
-                .iter()
-                .map(|value| value.to_string())
-                .collect(),
-        );
-    }
-
-    None
-}
-
-fn model_extension(model_path: &Path) -> String {
-    model_path
-        .extension()
-        .and_then(|value| value.to_str())
-        .unwrap_or_default()
-        .to_ascii_lowercase()
-}
-
-fn is_pytorch_checkpoint(model_path: &Path) -> bool {
-    matches!(model_extension(model_path).as_str(), "pt" | "pth")
 }
 
 fn trim_command_detail(detail: &str) -> String {
@@ -1317,38 +1145,6 @@ fn prepare_model_asset_for_inference(model_path: &Path) -> PreparedModelAsset {
             })),
         },
     }
-}
-
-/// Ultralytics-style filename suffix → task, authoritative over the installed
-/// category (a `-cls`/`-seg`/`-pose` asset is that task even if it was downloaded
-/// under a "detection" catalog entry). Returns `None` for plain detectors.
-fn task_suffix(model_path: &Path) -> Option<&'static str> {
-    let name = model_path
-        .file_name()
-        .and_then(|value| value.to_str())
-        .unwrap_or_default()
-        .to_lowercase();
-    if name.contains("-cls") || name.contains("_cls") {
-        Some("classification")
-    } else if name.contains("-pose") || name.contains("_pose") {
-        Some("pose")
-    } else if name.contains("-seg") || name.contains("_seg") {
-        Some("segmentation")
-    } else {
-        None
-    }
-}
-
-/// Accurate, role-aware reason a model can't run the box detector — segmentation
-/// models are usable (via the copilot), classification/pose are different tasks.
-fn non_detection_reason(task: &str) -> String {
-    match task {
-        "segmentation" => "Segmentation model \u{2014} used by the AI Copilot for click/box \u{2192} polygon (ask it to \u{201c}outline them\u{201d}), not the Detect button.",
-        "classification" => "Classification model \u{2014} it labels the whole image, not regions. Install a detection model (e.g. YOLO26 Detection) to draw boxes.",
-        "pose" => "Pose-estimation model \u{2014} not a box detector. Install a detection model for AI detect.",
-        _ => "AI detect supports object-detection models; this model uses a different task.",
-    }
-    .to_string()
 }
 
 fn prediction_unsupported_reason(
@@ -1664,14 +1460,6 @@ fn draft_matches_class(draft: &InferenceAnnotationDraft, target: &str) -> bool {
     .any(|value| class_token_matches(value.trim(), &target))
 }
 
-/// Singular/plural-tolerant equality for two already-lowercased class tokens.
-fn class_token_matches(a: &str, b: &str) -> bool {
-    if a.is_empty() || b.is_empty() {
-        return false;
-    }
-    a == b || format!("{a}s") == b || format!("{b}s") == a
-}
-
 /// Best-effort mapping from an installed `ai_models` entity to a registry plugin
 /// id (see [`crate::domain::ai::registry`]), used when the caller didn't pass an
 /// explicit `registryId`. The frontend normally passes the id of the catalog
@@ -1721,104 +1509,6 @@ fn registry_id_for_model(model: &Value) -> String {
     } else {
         "unknown".to_string()
     }
-}
-
-fn infer_model_family_and_variant(name: &str, model_path: &Path) -> (String, String) {
-    let haystack = format!(
-        "{} {}",
-        name.to_ascii_lowercase(),
-        model_path
-            .file_name()
-            .and_then(|value| value.to_str())
-            .unwrap_or_default()
-            .to_ascii_lowercase()
-    );
-
-    let family = if haystack.contains("yoloe-26") {
-        "yoloe-26"
-    } else if haystack.contains("yolo26") {
-        "yolo26"
-    } else if haystack.contains("yolo11") {
-        "yolo11"
-    } else if haystack.contains("yolov8") {
-        "yolov8"
-    } else {
-        ""
-    };
-
-    let variant = ["n", "s", "m", "l", "x"]
-        .iter()
-        .copied()
-        .find(|variant| {
-            haystack.contains(&format!("yolo26{variant}"))
-                || haystack.contains(&format!("yoloe-26{variant}"))
-        })
-        .unwrap_or_default();
-
-    (family.to_string(), variant.to_string())
-}
-
-fn infer_default_rank(family: &str, category: &str, variant: &str) -> i64 {
-    if family == "yolo26" && category == "detection" {
-        match variant {
-            "n" => 0,
-            "s" => 10,
-            "m" => 20,
-            "l" => 30,
-            "x" => 40,
-            _ => 100,
-        }
-    } else if family == "yolo26" {
-        50
-    } else if family == "yoloe-26" {
-        100
-    } else {
-        999
-    }
-}
-
-fn infer_model_runtime(model_path: &Path) -> (&'static str, &'static str) {
-    match model_extension(model_path).as_str() {
-        "onnx" => ("onnx", "ort"),
-        "pt" | "pth" => ("pytorch", "cpu"),
-        "tflite" => ("tflite", "cpu"),
-        "h5" => ("keras", "cpu"),
-        "pb" => ("tensorflow", "cpu"),
-        _ => ("unknown", "cpu"),
-    }
-}
-
-fn task_type_for_category(category: &str) -> &'static str {
-    match category {
-        "segmentation" => "segmentation",
-        "classification" => "classification",
-        "pose" => "pose_estimation",
-        "tracking" => "tracking",
-        _ => "object_detection",
-    }
-}
-
-fn build_model_version(
-    name: &str,
-    version: &str,
-    family: &str,
-    variant: &str,
-    category: &str,
-) -> String {
-    if !family.is_empty() && !variant.is_empty() {
-        let base = if family == "yoloe-26" {
-            format!("YOLOE-26{variant}")
-        } else {
-            format!("YOLO26{variant}")
-        };
-
-        return match category {
-            "segmentation" => format!("{base}-seg"),
-            "pose" => format!("{base}-pose"),
-            _ => base,
-        };
-    }
-    format!("{name} {version}")
 }
 
 fn file_name_from_path(path: &Path, invalid_message: &str) -> Result<String, AppError> {
